refactor CommitMessages & unit tests
