--- src/bot_actions/commit_messages/commit_message_note.ts
@@ -1,7 +1,0 @@
-/**
- * This class extends the `BotActionNote` class by analyzing different state combinations unique to the Commit Messages Action.
- * Each instance of this class contains a message string that provides feedback to the end-user about the titles of the commits contained in the GitLab Merge Request.
-//  */
-// export class CommitMessagesNote {
-
-// }
